Added File Dialog and Portable Mod Groundwork
Changed the user file input from a text field to a QFile dialog, as it should've been from the start.

Also added the beginning of work for mod files that can be easily moved from one user to another. Conventions for mod files are described in modhandler.cpp, and will be finished when the user input for address ranges is figured out.