Added Settings Menu, Changed to ListWidget, Bugfixes
-Added Settings Menu
Finally added a settings menu. Current settings are the file's address offset, the browser's starting address, starting window size (default recommended), and how much the scroll buttons move the code browser.

-Changed to ListWidget
Previously, the code browser was written as a QPlainTextEdit. This was, for many reasons, a bad idea. It's now a QListWidget, which should be more robust and easier to work with.

-Bugfixes
A lot of them over the past couple days. Didn't keep track of them, but they're fixed now.